move Debuger into package tools, update documnetation
